Cambiada la funcionalidad al módulo census
