+++ decide/authentication/templates/master.html
@@ -49,7 +49,7 @@
                         </ul>
                     </div>
                 </li>
-                <li><a href='/authentication/userVotings'>Mis votaciones</a></li>
+                <li><a href='/census/userVotings'>Mis votaciones</a></li>
                 <li><a href="/census/operations">Operaciones entre grupos</a></li>
 
                 {%else %}
